Choice of continuation in case expression is now done via method in the closure containing the case expression, not in the closure of the constructor.
I did this change as i did not understand how a constructor closure could chose the continuation
without array_key_exists-calls to the previous return vector. I also need a hook for updates in
another branch. So i now let the case closure chose a continuation witin itself and the
constructor/primitive return their content to the case-closure. This in turn is done via a
conventional array in the arguments register.