Tests run again, with something i consider an update.
There are no tests whether the update stuff really works. There are also no tests on updates
of partial function. There also is a really weird method for merging two stacks, as SPLStack
seems to have no merge function and there is confusion about the order given through the Array-
and Iterator-Interface of SPLStack. As we introduced indirections for updates, we also need
some garbage collection to get rid of indirections. The hierarchy for Closures is broken also,
as there clearly exist closures that need no free variables.